feat: update quests copy writing and file naming
