+++ mobile-app/lib/features/main/screens/quests/king_of_the_shill_screen.dart
@@ -1,12 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:quantus_sdk/quantus_sdk.dart';
-import 'package:resonance_network_wallet/features/components/basic_card.dart';
 import 'package:resonance_network_wallet/features/components/button.dart';
-import 'package:resonance_network_wallet/features/components/copy_icon.dart';
 import 'package:resonance_network_wallet/features/components/link_text.dart';
+import 'package:resonance_network_wallet/features/components/raid_submission_action_sheet.dart';
 import 'package:resonance_network_wallet/features/components/scaffold_base.dart';
-import 'package:resonance_network_wallet/features/components/skeleton.dart';
 import 'package:resonance_network_wallet/features/components/sphere.dart';
 import 'package:resonance_network_wallet/features/components/wallet_app_bar.dart';
 import 'package:resonance_network_wallet/features/main/screens/quests/account_associations_status.dart';
@@ -14,12 +12,9 @@ import 'package:resonance_network_wallet/features/main/screens/quests/optin_posi
 import 'package:resonance_network_wallet/features/main/screens/quests/quest_title.dart';
 import 'package:resonance_network_wallet/features/styles/app_colors_theme.dart';
 import 'package:resonance_network_wallet/features/styles/app_text_theme.dart';
-import 'package:resonance_network_wallet/providers/account_associations_providers.dart';
-import 'package:resonance_network_wallet/providers/account_stats_providers.dart';
-import 'package:resonance_network_wallet/services/referral_service.dart';
-import 'package:resonance_network_wallet/shared/extensions/clipboard_extensions.dart';
+import 'package:resonance_network_wallet/providers/raider_quest_providers.dart';
 import 'package:resonance_network_wallet/shared/extensions/media_query_data_extension.dart';
-import 'package:share_plus/share_plus.dart';
+import 'package:resonance_network_wallet/shared/extensions/snackbar_extensions.dart';
 
 class KingOfTheShillScreen extends ConsumerStatefulWidget {
   const KingOfTheShillScreen({super.key});
@@ -29,40 +24,12 @@ class KingOfTheShillScreen extends ConsumerStatefulWidget {
 }
 
 class _KingOfTheShillScreenState extends ConsumerState<KingOfTheShillScreen> with WidgetsBindingObserver {
-  final ReferralService _referralService = ReferralService();
+  final _taskmasterService = TaskmasterService();
   final ScrollController _scrollController = ScrollController();
-
-  String? _referralCode;
-
-  Future<void> _loadReferralCode() async {
-    try {
-      final myReferralCode = await _referralService.getMyInviteCode();
-      setState(() {
-        _referralCode = myReferralCode;
-      });
-    } catch (e) {
-      debugPrint('Error loading account data: $e');
-      if (mounted) {
-        setState(() {});
-      }
-    }
-  }
-
-  Future<void> _shareReferralLink() async {
-    final params = await _referralService.getShareLinkParameters(context.sharePositionRect());
-    SharePlus.instance.share(params);
-  }
-
-  void _copyReferralCode() {
-    if (_referralCode != null) {
-      ClipboardExtensions.copyTextWithSnackbar(context, _referralCode!, message: 'Referral code copied to clipboard');
-    }
-  }
 
   @override
   void initState() {
     super.initState();
-    _loadReferralCode();
   }
 
   @override
@@ -70,24 +37,50 @@ class _KingOfTheShillScreenState extends ConsumerState<KingOfTheShillScreen> wit
     super.dispose();
   }
 
-  void refreshStatsData() {
-    ref.invalidate(accountsStatsProvider);
+  void refreshRaiderSubmissions() {
+    ref.invalidate(raiderSubmissionsProvider);
   }
 
-  void refreshAssociationsData() {
-    ref.invalidate(accountAssociationsProvider);
+  String? extractXStatusId(String url) {
+    final uri = Uri.tryParse(url);
+    if (uri == null) return null;
+
+    // Expected path: /{username}/status/{id}
+    final segments = uri.pathSegments;
+
+    if (segments.length >= 3 && segments[1] == 'status') {
+      final id = segments[2];
+      return RegExp(r'^\d+$').hasMatch(id) ? id : null;
+    }
+
+    return null;
+  }
+
+  Future<void> _handleRemoveSubmission(String id) async {
+    try {
+      await _taskmasterService.removeRaidSubmission(id);
+      if (mounted) {
+        context.showSuccessSnackbar(title: 'Success removed!', message: 'Success removing raid submission!');
+      }
+      ref.invalidate(raiderSubmissionsProvider);
+    } catch (e) {
+      print('Failed removing raid submission: $e');
+
+      if (mounted) {
+        context.showErrorSnackbar(title: 'Failed removing!', message: e.toString());
+      }
+    }
   }
 
   @override
   Widget build(BuildContext context) {
-    final statsAsync = ref.watch(accountsStatsProvider);
+    final raiderSubmissionsAsync = ref.watch(raiderSubmissionsProvider);
+    final effectiveSpacing = context.isSmallHeight ? 24.0 : 36.0;
 
     return ScaffoldBase.refreshable(
       appBar: WalletAppBar(title: 'King of The Shill'),
-      padding: const EdgeInsetsGeometry.all(0),
       onRefresh: () async {
-        refreshStatsData();
-        refreshAssociationsData();
+        refreshRaiderSubmissions();
       },
       scrollController: _scrollController,
       decorations: [
@@ -102,74 +95,124 @@ class _KingOfTheShillScreenState extends ConsumerState<KingOfTheShillScreen> wit
                 padding: const EdgeInsets.only(top: 11),
                 child: Column(
                   children: [
-                    Padding(
-                      padding: const EdgeInsets.only(right: 44),
-                      child: Row(
-                        crossAxisAlignment: CrossAxisAlignment.start,
-                        spacing: 17,
-                        children: [
-                          _buildDecoration(),
-                          Expanded(
-                            child: Column(
-                              crossAxisAlignment: CrossAxisAlignment.start,
-                              children: [
-                                const SizedBox(height: 96),
-                                const Row(
-                                  crossAxisAlignment: CrossAxisAlignment.center,
-                                  mainAxisAlignment: MainAxisAlignment.center,
-                                  children: [OptinPositionStatus(), SizedBox(width: 71)],
-                                ),
-                                SizedBox(height: context.isSmallHeight ? 18 : 37.0),
-                                const AccountAssociationsStatus(),
-                                SizedBox(height: context.isSmallHeight ? 18 : 37.0),
-                                ..._buildAccountStats(context, statsAsync),
-                                const SizedBox(height: 16),
-                                LinkText(
-                                  label: 'Learn more about QQ',
-                                  url: AppConstants.shillQuestsPageUrl,
-                                  textStyle: context.themeText.smallParagraph,
-                                ),
-                              ],
-                            ),
-                          ),
-                        ],
-                      ),
-                    ),
-                    SizedBox(height: context.isSmallHeight ? 18 : 40),
-                    Padding(
-                      padding: const EdgeInsets.symmetric(horizontal: 24.0),
-                      child: InkWell(
-                        onTap: _copyReferralCode,
-                        child: Container(
-                          padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 11),
-                          decoration: ShapeDecoration(
-                            color: context.themeColors.background,
-                            shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(4)),
-                          ),
-                          child: Row(
-                            spacing: 12.0,
-                            mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                    Row(
+                      crossAxisAlignment: CrossAxisAlignment.start,
+                      spacing: 17,
+                      children: [
+                        Expanded(
+                          child: Column(
+                            crossAxisAlignment: CrossAxisAlignment.start,
                             children: [
-                              Text(
-                                _referralCode ?? 'Loading...',
-                                style: context.themeText.smallParagraph,
-                                textAlign: TextAlign.center,
+                              const SizedBox(height: 96),
+                              const Row(
+                                crossAxisAlignment: CrossAxisAlignment.center,
+                                mainAxisAlignment: MainAxisAlignment.center,
+                                children: [OptinPositionStatus()],
                               ),
-                              const CopyIcon(),
+                              SizedBox(height: effectiveSpacing),
+                              const AccountAssociationsStatus(),
+                              SizedBox(height: effectiveSpacing),
+                              ...raiderSubmissionsAsync.when(
+                                loading: () => [
+                                  Center(child: CircularProgressIndicator(color: context.themeColors.circularLoader)),
+                                ],
+                                error: (error, stackTrace) => [
+                                  Text(
+                                    'Error fetching raider submissions.',
+                                    style: context.themeText.detail?.copyWith(color: context.themeColors.textError),
+                                  ),
+                                  const SizedBox(height: 12),
+                                  Button(
+                                    variant: ButtonVariant.neutral,
+                                    label: 'Try again',
+                                    onPressed: refreshRaiderSubmissions,
+                                  ),
+                                ],
+                                data: (state) {
+                                  switch (state) {
+                                    case RaiderSubmissionsOk():
+                                      return [
+                                        LinkText(
+                                          label: 'Learn more',
+                                          url: AppConstants.raidQuestsPageUrl,
+                                          textStyle: context.themeText.smallParagraph,
+                                        ),
+                                        SizedBox(height: effectiveSpacing),
+                                        Row(
+                                          mainAxisSize: MainAxisSize.min,
+                                          children: [
+                                            Text('Raid Submissions', style: context.themeText.smallTitle),
+                                            const SizedBox(width: 12),
+                                            InkWell(
+                                              child: Container(
+                                                decoration: ShapeDecoration(
+                                                  color: context.themeColors.buttonNeutral,
+                                                  shape: RoundedRectangleBorder(
+                                                    borderRadius: BorderRadiusGeometry.circular(8),
+                                                  ),
+                                                ),
+                                                child: const Icon(Icons.add, color: Colors.black),
+                                              ),
+                                              onTap: () {
+                                                showRaidSubmissionActionSheet(context);
+                                              },
+                                            ),
+                                          ],
+                                        ),
+                                        const SizedBox(height: 8),
+                                        if (state.submissions.isNotEmpty)
+                                          Column(
+                                            spacing: 4,
+                                            children: state.submissions.asMap().entries.map((entry) {
+                                              final index = entry.key + 1;
+                                              final value = entry.value;
+                                              final label = extractXStatusId(value) ?? 'Unknown';
+
+                                              return Row(
+                                                children: [
+                                                  Text('$index. '),
+                                                  LinkText(
+                                                    label: label,
+                                                    url: value,
+                                                    textStyle: context.themeText.smallParagraph,
+                                                  ),
+                                                  InkWell(
+                                                    child: Icon(Icons.delete, color: context.themeColors.buttonDanger),
+                                                    onTap: () {
+                                                      _handleRemoveSubmission(label);
+                                                    },
+                                                  ),
+                                                ],
+                                              );
+                                            }).toList(),
+                                          )
+                                        else
+                                          Text(
+                                            "You haven't submitted anything yet",
+                                            style: context.themeText.smallParagraph,
+                                          ),
+                                      ];
+
+                                    case NoActiveRaid():
+                                      return [
+                                        LinkText(
+                                          label: 'Learn more',
+                                          url: AppConstants.raidQuestsPageUrl,
+                                          textStyle: context.themeText.smallParagraph,
+                                        ),
+                                      ];
+
+                                    case NoTwitterLinked():
+                                      return [Text('Please link your X account', style: context.themeText.smallTitle)];
+                                  }
+                                },
+                              ),
                             ],
                           ),
                         ),
-                      ),
+                      ],
                     ),
-                    const SizedBox(height: 16),
-                    Padding(
-                      padding: const EdgeInsets.symmetric(horizontal: 24),
-                      child: Button(
-                        variant: ButtonVariant.glassOutline,
-                        label: 'Share Referral Link',
-                        onPressed: _shareReferralLink,
-                      ),
-                    ),
+                    SizedBox(height: context.isSmallHeight ? 18 : 40),
                   ],
                 ),
               ),
@@ -178,68 +221,6 @@ class _KingOfTheShillScreenState extends ConsumerState<KingOfTheShillScreen> wit
           ),
         ),
       ],
-    );
-  }
-
-  List<Widget> _buildAccountStats(BuildContext context, AsyncValue<AccountStats> statsAsync) {
-    return statsAsync.when(
-      data: (stats) => [
-        _buildStatCard(context, 'Referrals:', stats.referralCount),
-        const SizedBox(height: 9),
-        _buildStatCard(context, 'Sends:', stats.sendCount),
-        const SizedBox(height: 9),
-        _buildStatCard(context, 'Reversals:', stats.reversalCount),
-        const SizedBox(height: 9),
-        _buildStatCard(context, 'Mining:', stats.miningCount),
-      ],
-      loading: () => [
-        _buildStatCard(context, 'Referrals:', null),
-        const SizedBox(height: 9),
-        _buildStatCard(context, 'Sends:', null),
-        const SizedBox(height: 9),
-        _buildStatCard(context, 'Reversals:', null),
-        const SizedBox(height: 9),
-        _buildStatCard(context, 'Mining:', null),
-      ],
-      error: (error, stack) => [
-        Text(
-          'Error fetching account stats.',
-          style: context.themeText.detail?.copyWith(color: context.themeColors.textError),
-        ),
-        const SizedBox(height: 12),
-        Button(variant: ButtonVariant.neutral, label: 'Try again', onPressed: refreshStatsData),
-      ],
-    );
-  }
-
-  Widget _buildStatCard(BuildContext context, String title, int? stat) {
-    final isLoading = stat == null;
-
-    return BasicCard(
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-        children: [
-          Text(title, style: context.themeText.smallTitle),
-          isLoading ? const Skeleton(width: 40, height: 16) : Text('$stat', style: context.themeText.smallTitle),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildDecoration() {
-    return Container(
-      width: 85,
-      height: context.isSmallHeight ? 415 : 480,
-      decoration: const ShapeDecoration(
-        gradient: LinearGradient(
-          begin: Alignment(0.03, -1.00),
-          end: Alignment(-0.03, 1),
-          colors: [Color(0xFF0000FF), Color(0xFFED4CCE), Color(0xFFFFE91F)],
-        ),
-        shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.only(topRight: Radius.circular(10), bottomRight: Radius.circular(10)),
-        ),
-      ),
     );
   }
 }
